migrated refactor code into lib, removed all debug prints
